Replace RecaptchaForm with RecaptchaEnterpriseForm (#12954) (#13057)
Replace RecaptchaForm with RecaptchaEnterpriseForm (#12954)

changelog: Internal, reCaptcha, Replace RecaptchaForm with RecaptchaEnterpriseForm

Fixes https://gsa-standard.atlassian-us-gov-mod.net/browse/LG-17458

- Make RecaptchaMockForm inherit from RecaptchaEnterpriseForm
- Make RecaptchaEnterpriseForm the default
- Remove RecaptchaForm
- Roll recaptcha_enterprise? into recaptcha_enabled?
- Removed unused config var `recaptcha_secret_key`
